VCST-1385: External link encoded after saving (#23)
fix: Resolves issue with External link encoded after saving

+++ src/VirtoCommerce.AzureBlobAssetsModule.Core/AzureBlobProvider.cs
@@ -1,13 +1,13 @@
+using Azure.Storage.Blobs;
+using Azure.Storage.Blobs.Models;
+using Azure.Storage.Blobs.Specialized;
+using Microsoft.Extensions.Options;
 using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
 using System.Web;
-using Azure.Storage.Blobs;
-using Azure.Storage.Blobs.Models;
-using Azure.Storage.Blobs.Specialized;
-using Microsoft.Extensions.Options;
 using VirtoCommerce.Assets.Abstractions;
 using VirtoCommerce.AssetsModule.Core.Assets;
 using VirtoCommerce.Platform.Core;
@@ -72,7 +72,7 @@ namespace VirtoCommerce.AzureBlobAssetsModule.Core
             {
                 var blob = GetBlockBlobClient(blobUrl);
                 var props = await blob.GetPropertiesAsync();
-                result = ConvertBlobToBlobInfo(blob, props.Value);
+                result = ConvertToBlobInfo(blob, props.Value);
             }
             catch
             {
@@ -153,16 +153,25 @@ namespace VirtoCommerce.AzureBlobAssetsModule.Core
 
         public virtual async Task RemoveAsync(string[] urls)
         {
-            foreach (var url in urls.Where(x => !string.IsNullOrWhiteSpace(x)))
+            ArgumentNullException.ThrowIfNull(urls);
+
+            var urlsToDelete = urls.Where(x => !string.IsNullOrWhiteSpace(x)).ToArray();
+
+            if (urlsToDelete.Length == 0)
             {
-                var absoluteUri = url.IsAbsoluteUrl()
-                    ? new Uri(url).ToString()
-                    : UrlHelperExtensions.Combine(_blobServiceClient.Uri.ToString(), url);
+                return;
+            }
+
+            foreach (var url in urlsToDelete)
+            {
+                var uri = GetAbsoluteUri(_blobServiceClient.Uri, url);
+                var absoluteUri = uri.AbsoluteUri;
                 var container = GetBlobContainerClient(absoluteUri);
 
                 var isFolder = string.IsNullOrEmpty(Path.GetFileName(absoluteUri));
-                var blobSearchPrefix = isFolder ? GetDirectoryPathFromUrl(absoluteUri)
-                                             : GetFilePathFromUrl(absoluteUri);
+                var blobSearchPrefix = isFolder
+                    ? GetDirectoryPathFromUrl(absoluteUri)
+                    : GetFilePathFromUrl(absoluteUri);
 
                 if (string.IsNullOrEmpty(blobSearchPrefix))
                 {
@@ -173,6 +182,10 @@ namespace VirtoCommerce.AzureBlobAssetsModule.Core
                     var blobItems = container.GetBlobsAsync(prefix: blobSearchPrefix);
                     await foreach (var blobItem in blobItems)
                     {
+                        if (blobItem.Name != blobSearchPrefix)
+                        {
+                            continue;
+                        }
                         var blobClient = container.GetBlobClient(blobItem.Name);
                         await blobClient.DeleteIfExistsAsync();
                     }
@@ -190,7 +203,7 @@ namespace VirtoCommerce.AzureBlobAssetsModule.Core
 
                 if (await container.ExistsAsync())
                 {
-                    var baseUriEscaped = container.Uri.AbsoluteUri; // absoluteUri is escaped already
+                    var baseUri = container.Uri; // absoluteUri is escaped already
                     var prefix = GetDirectoryPathFromUrl(folderUrl);
                     if (!string.IsNullOrEmpty(keyword))
                     {
@@ -201,8 +214,7 @@ namespace VirtoCommerce.AzureBlobAssetsModule.Core
                     var containerProperties = await container.GetPropertiesAsync();
 
                     // Call the listing operation and return pages of the specified size.
-                    var resultSegment = container.GetBlobsByHierarchyAsync(prefix: prefix, delimiter: Delimiter)
-                        .AsPages();
+                    var resultSegment = container.GetBlobsByHierarchyAsync(prefix: prefix, delimiter: Delimiter).AsPages();
 
                     // Enumerate the blobs returned for each page.
                     await foreach (var blobPage in resultSegment)
@@ -212,19 +224,12 @@ namespace VirtoCommerce.AzureBlobAssetsModule.Core
                         {
                             if (blobHierarchyItem.IsPrefix)
                             {
-                                var folder = AbstractTypeFactory<BlobFolder>.TryCreateInstance();
-
-                                folder.Name = GetOutlineFromUrl(blobHierarchyItem.Prefix).Last();
-                                folder.Url = UrlHelperExtensions.Combine(baseUriEscaped, EscapeUri(blobHierarchyItem.Prefix));
-                                folder.ParentUrl = GetParentUrl(baseUriEscaped, blobHierarchyItem.Prefix);
-                                folder.RelativeUrl = folder.Url.Replace(_blobServiceClient.Uri.AbsoluteUri.TrimEnd(Delimiter[0]), string.Empty);
-                                folder.CreatedDate = containerProperties.Value.LastModified.UtcDateTime;
-                                folder.ModifiedDate = containerProperties.Value.LastModified.UtcDateTime;
+                                var folder = ConvertToBlobFolder(blobHierarchyItem, baseUri, containerProperties.Value);
                                 result.Results.Add(folder);
                             }
                             else
                             {
-                                var blobInfo = ConvertBlobToBlobInfo(blobHierarchyItem.Blob, baseUriEscaped);
+                                var blobInfo = ConvertToBlobInfo(blobHierarchyItem.Blob, baseUri);
                                 //Do not return empty blob (created with directory because azure blob not support direct directory creation)
                                 if (!string.IsNullOrEmpty(blobInfo.Name))
                                 {
@@ -242,13 +247,7 @@ namespace VirtoCommerce.AzureBlobAssetsModule.Core
 
                 await foreach (var containerPage in resultSegment)
                 {
-                    var folders = containerPage.Values.Select(x =>
-                    {
-                        var folder = AbstractTypeFactory<BlobFolder>.TryCreateInstance();
-                        folder.Name = x.Name.Split(Delimiter).Last();
-                        folder.Url = EscapeUri(UrlHelperExtensions.Combine(_blobServiceClient.Uri.ToString(), x.Name));
-                        return folder;
-                    });
+                    var folders = containerPage.Values.Select(ConvertToBlobFolder);
                     result.Results.AddRange(folders);
                 }
             }
@@ -259,9 +258,14 @@ namespace VirtoCommerce.AzureBlobAssetsModule.Core
 
         public virtual async Task CreateFolderAsync(BlobFolder folder)
         {
-            var path = folder.ParentUrl == null ?
-                        folder.Name :
-                        UrlHelperExtensions.Combine(folder.ParentUrl, folder.Name);
+            var path = folder.Name;
+            if (!string.IsNullOrEmpty(folder.ParentUrl))
+            {
+                var parentUrl = folder.ParentUrl.IsAbsoluteUrl()
+                    ? GetRelativeUrl(new Uri(folder.ParentUrl))
+                    : folder.ParentUrl;
+                path = UrlHelperExtensions.Combine(parentUrl, folder.Name);
+            }
 
             var container = await CreateContainerIfNotExists(path);
 
@@ -372,23 +376,18 @@ namespace VirtoCommerce.AzureBlobAssetsModule.Core
 
         #region IBlobUrlResolver Members
 
-        public string GetAbsoluteUrl(string blobKey)
+        public string GetAbsoluteUrl(string inputUrl)
         {
-            var result = blobKey;
-            if (!blobKey.IsAbsoluteUrl())
+            ArgumentNullException.ThrowIfNull(nameof(inputUrl));
+
+            var baseUri = _blobServiceClient.Uri;
+            if (!string.IsNullOrWhiteSpace(_cdnUrl))
             {
-                var baseUrl = _blobServiceClient.Uri.AbsoluteUri;
-
-                if (!string.IsNullOrWhiteSpace(_cdnUrl))
-                {
-                    var cdnUriBuilder = new UriBuilder(_blobServiceClient.Uri.Scheme, _cdnUrl);
-                    baseUrl = cdnUriBuilder.Uri.AbsoluteUri;
-                }
-
-                result = UrlHelperExtensions.Combine(baseUrl, EscapeUri(blobKey));
+                var cdnUriBuilder = new UriBuilder(_blobServiceClient.Uri.Scheme, _cdnUrl);
+                baseUri = cdnUriBuilder.Uri;
             }
 
-            return result;
+            return GetAbsoluteUri(baseUri, inputUrl).AbsoluteUri;
         }
 
         #endregion IBlobUrlResolver Members
@@ -411,7 +410,7 @@ namespace VirtoCommerce.AzureBlobAssetsModule.Core
         /// </summary>
         /// <param name="url"></param>
         /// <returns></returns>
-        private string[] GetOutlineFromUrl(string url)
+        private static string[] GetOutlineFromUrl(string url)
         {
             var relativeUrl = url;
             if (url.IsAbsoluteUrl())
@@ -434,40 +433,21 @@ namespace VirtoCommerce.AzureBlobAssetsModule.Core
             return relativeUrl.Split(Delimiter[0], '\\'); // name may be empty
         }
 
-        private string GetContainerNameFromUrl(string url)
+        private static string GetContainerNameFromUrl(string url)
         {
             return GetOutlineFromUrl(url).First();
         }
 
-        private string GetDirectoryPathFromUrl(string url)
+        private static string GetDirectoryPathFromUrl(string url)
         {
             var result = string.Join(Delimiter, GetOutlineFromUrl(url).Skip(1).ToArray());
             return !string.IsNullOrEmpty(result) ? HttpUtility.UrlDecode(result) + Delimiter : null;
         }
 
-        private string GetFilePathFromUrl(string url)
+        private static string GetFilePathFromUrl(string url)
         {
             var result = string.Join(Delimiter, GetOutlineFromUrl(url).Skip(1).ToArray());
             return !string.IsNullOrEmpty(result) ? HttpUtility.UrlDecode(result) : null;
-        }
-
-        private string GetParentUrl(string baseUri, string blobPrefix)
-        {
-            var segments = GetOutlineFromUrl(blobPrefix);
-            var parentPath = string.Join(Delimiter, segments.Take(segments.Length - 1));
-            return UrlHelperExtensions.Combine(baseUri, EscapeUri(parentPath));
-        }
-
-        private static string EscapeUri(string stringToEscape)
-        {
-            var uri = new Uri(stringToEscape, UriKind.RelativeOrAbsolute);
-            if (uri.IsAbsoluteUri)
-            {
-                return uri.AbsoluteUri;
-            }
-
-            var parts = stringToEscape.Split(Delimiter[0]);
-            return string.Join(Delimiter, parts.Select(Uri.EscapeDataString));
         }
 
         private BlockBlobClient GetBlockBlobClient(string blobUrl)
@@ -487,42 +467,153 @@ namespace VirtoCommerce.AzureBlobAssetsModule.Core
             return container;
         }
 
-        private BlobInfo ConvertBlobToBlobInfo(BlobBaseClient blob, BlobProperties props)
+        private BlobInfo ConvertToBlobInfo(BlobBaseClient blob, BlobProperties props)
         {
-            var absoluteUrl = blob.Uri;
-            var relativeUrl = Delimiter + UrlHelperExtensions.Combine(GetContainerNameFromUrl(blob.Uri.AbsoluteUri), EscapeUri(blob.Name));
-            var fileName = Path.GetFileName(Uri.UnescapeDataString(blob.Name));
-            var contentType = MimeTypeResolver.ResolveContentType(fileName);
+            var blobInfo = AbstractTypeFactory<BlobInfo>.TryCreateInstance();
 
-            return new BlobInfo
-            {
-                Url = absoluteUrl.AbsoluteUri,
-                Name = fileName,
-                ContentType = contentType,
-                Size = props.ContentLength,
-                CreatedDate = props.CreatedOn.UtcDateTime,
-                ModifiedDate = props.LastModified.UtcDateTime,
-                RelativeUrl = relativeUrl
-            };
+            var absoluteUri = blob.Uri;
+            var relativeUrl = GetRelativeUrl(absoluteUri);
+            var fileName = Path.GetFileName(Uri.UnescapeDataString(blob.Name));
+
+            blobInfo.Url = absoluteUri.AbsoluteUri;
+            blobInfo.RelativeUrl = relativeUrl;
+            blobInfo.Name = fileName;
+            blobInfo.ContentType = props.ContentType.EmptyToNull() ?? MimeTypeResolver.ResolveContentType(fileName);
+            blobInfo.Size = props.ContentLength;
+            blobInfo.CreatedDate = props.CreatedOn.UtcDateTime;
+            blobInfo.ModifiedDate = props.LastModified.UtcDateTime;
+
+            return blobInfo;
         }
 
-        private BlobInfo ConvertBlobToBlobInfo(BlobItem blob, string baseUri)
+        private BlobInfo ConvertToBlobInfo(BlobItem blob, Uri baseUri)
         {
-            var fileName = Path.GetFileName(blob.Name);
-            var absoluteUrl = UrlHelperExtensions.Combine(baseUri, EscapeUri(blob.Name));
-            var relativeUrl = Delimiter + absoluteUrl.Replace(EscapeUri(_blobServiceClient.Uri.ToString()), string.Empty);
-            var contentType = MimeTypeResolver.ResolveContentType(fileName);
+            var blobInfo = AbstractTypeFactory<BlobInfo>.TryCreateInstance();
 
-            return new BlobInfo
+            var absoluteUri = GetAbsoluteUri(baseUri, blob.Name);
+            var relativeUrl = GetRelativeUrl(absoluteUri);
+            var fileName = Path.GetFileName(Uri.UnescapeDataString(blob.Name));
+
+            blobInfo.Url = absoluteUri.AbsoluteUri;
+            blobInfo.RelativeUrl = relativeUrl;
+            blobInfo.Name = fileName;
+            blobInfo.ContentType = blob.Properties.ContentType.EmptyToNull() ?? MimeTypeResolver.ResolveContentType(fileName);
+            blobInfo.Size = blob.Properties.ContentLength ?? 0;
+            blobInfo.CreatedDate = blob.Properties.CreatedOn!.Value.UtcDateTime;
+            blobInfo.ModifiedDate = blob.Properties.LastModified?.UtcDateTime;
+
+            return blobInfo;
+        }
+
+        private BlobFolder ConvertToBlobFolder(BlobHierarchyItem blobHierarchyItem, Uri baseUri, BlobContainerProperties containerProperties)
+        {
+            var folder = AbstractTypeFactory<BlobFolder>.TryCreateInstance();
+
+            var absoluteUri = GetAbsoluteUri(baseUri, blobHierarchyItem.Prefix);
+
+            folder.Name = GetOutlineFromUrl(blobHierarchyItem.Prefix).Last();
+            folder.Url = absoluteUri.AbsoluteUri;
+            folder.ParentUrl = GetParentUrl(baseUri, blobHierarchyItem.Prefix);
+            folder.RelativeUrl = GetRelativeUrl(absoluteUri);
+            folder.ModifiedDate = folder.CreatedDate = containerProperties.LastModified.UtcDateTime;
+
+            return folder;
+        }
+
+        private BlobFolder ConvertToBlobFolder(BlobContainerItem container)
+        {
+            var folder = AbstractTypeFactory<BlobFolder>.TryCreateInstance();
+
+            var baseUri = _blobServiceClient.Uri;
+            var absoluteUri = GetAbsoluteUri(baseUri, container.Name);
+
+            folder.Name = container.Name.Split(Delimiter).Last();
+            folder.Url = absoluteUri.AbsoluteUri;
+            folder.RelativeUrl = GetRelativeUrl(absoluteUri);
+            folder.ModifiedDate = folder.CreatedDate = container.Properties.LastModified.UtcDateTime;
+
+            return folder;
+        }
+
+        private static string GetParentUrl(Uri baseUri, string blobPrefix)
+        {
+            var baseUriString = baseUri.GetLeftPart(UriPartial.Path);
+            var baseUriQuery = baseUri.Query;
+            return GetParentUrl(baseUriString, blobPrefix, baseUriQuery);
+        }
+
+        private static string GetParentUrl(string baseUri, string blobPrefix, string query = null)
+        {
+            var segments = GetOutlineFromUrl(blobPrefix);
+            var parentPath = string.Join(Delimiter, segments.Take(segments.Length - 1));
+            return GetAbsoluteUri(new Uri(baseUri), parentPath, query).AbsoluteUri;
+        }
+
+        private string GetRelativeUrl(Uri absoluteUri)
+        {
+            if (!string.IsNullOrEmpty(absoluteUri.Query))
             {
-                Url = absoluteUrl,
-                Name = fileName,
-                ContentType = contentType,
-                Size = blob.Properties.ContentLength ?? 0,
-                CreatedDate = blob.Properties.CreatedOn!.Value.UtcDateTime,
-                ModifiedDate = blob.Properties.LastModified?.UtcDateTime,
-                RelativeUrl = relativeUrl
-            };
+                absoluteUri = new Uri(absoluteUri.GetLeftPart(UriPartial.Path));
+            }
+            var result = _blobServiceClient.Uri.MakeRelativeUri(absoluteUri).ToString();
+
+            // Ensure the relative URL starts with a delimiter
+            if (!result.StartsWith(Delimiter))
+            {
+                result = Delimiter + result;
+            }
+
+            return result;
+        }
+
+        private static Uri GetAbsoluteUri(Uri baseUri, string inputUrl, string query)
+        {
+            var result = GetAbsoluteUri(baseUri, inputUrl);
+
+            if (!string.IsNullOrEmpty(query))
+            {
+                if (string.IsNullOrEmpty(query))
+                {
+                    return result;
+                }
+
+                var builder = new UriBuilder(result)
+                {
+                    Query = query,
+                };
+                return builder.Uri;
+            }
+
+            return result;
+        }
+
+        private static Uri GetAbsoluteUri(Uri baseUri, string inputUrl)
+        {
+            ArgumentNullException.ThrowIfNull(nameof(inputUrl));
+
+            // do trim lead slash to prevent transform it to absolute file path on linux.
+            if (Uri.TryCreate(inputUrl.TrimStart('/'), UriKind.Absolute, out var resultUri))
+            {
+                // If the input URL is already absolute, return it as is (with correct encoding)
+                return resultUri;
+            }
+
+            if (inputUrl.StartsWith('/'))
+            {
+                inputUrl = "." + inputUrl;
+            }
+            else if (!inputUrl.StartsWith('.'))
+            {
+                inputUrl = "./" + inputUrl;
+            }
+
+            if (Uri.TryCreate(baseUri, inputUrl, out resultUri))
+            {
+                // If the input URL is relative, combine it with the base URI
+                return resultUri;
+            }
+
+            return new Uri(inputUrl);
         }
     }
 }
